Moved ouroboros-network-protocols to ouroboros-network
* `ouroboros-network-protocols`         -> `ouroboros-network:protocols` in `protocols/lib`
* `ouroboros-network-protocols:testlib` -> `ouroboros-network:protocols-tests-lib` in `protocols/tests-lib`
* `ouroboros-network-protocols:test`    -> `ouroboros-network:protocols-tests` in `protocols/tests`
* `ouroboros-network-protocols:bench`   -> `ouroboros-network:protocols-bench` in `protocols/bench`

+++ ouroboros-network/protocols/tests-lib/Test/ChainProducerState.hs
@@ -193,7 +193,7 @@ genFollowerState n c = do
     followerPoint <- frequency
       [ (2, return (headPoint c))
       , (2, return (mkRollbackPoint c n))
-      , (8, mkRollbackPoint c <$> choose (1, fromIntegral n - 1))
+      , (8, mkRollbackPoint c <$> choose (1, n - 1))
       ]
     followerNext <- oneof
       [ return FollowerForwardFrom
